feat: Implement test code from an example GraphicsView
To make sure that the animation and acceleration worked correctly, I
coded all of it in a fork of Assignment 6 and used it as a way to make
sure that the ball could speed up and slow down when the user swiped
the Ball.

The way it calculates is is by having a set max travel time, which is
used to divide the change in speed.

I also have checks to see that the user is swiping the ball and move it
if it does.