refactor: Give variables and methods more appropriate names and simplify code
All variables that affect the x and y axis have been standardized to have
the naming style <NAME>[x or y].

I changed animateCircle() to direction(), as the purpose of this method
is really to help determine which way the ball should roll. I removed
speed as a parameter and multiply the velocity with the direction
modifier within onDraw() where the movement is applied.

I have also removed some of the getters, such as getCollidedX() and
getCollidedY() as it did not make sense for outer classes to access it.
Instead, I have a new set method, getGesture(), which resets all the
important variables such as collided and the travelTime variable.

I have also added a movesLeft variable, which decrements everytime the
user swipes the screen.