feat: Increase the score every time the user hits the Target and decrease moves
In GraphicsView onDraw() method, the target constantly check if the ball's x
and y position is has intersected with the target. Every time it does, the
ball's position will reset and the screen will draw a new target to hit.

This increases the score and is displayed at the top of the screen. Once the
ball stops moving, the number of moves decrease and updates the number of
moves left at the top.