fix: Remove ambiguity of swipe direction by changing logic of direction()
It the logic that checks to see if the ball should go up or down, I had the
conditions set to >= and <=. This conflicts with the boolean check ==, which
makes sure if the ball is stationary along a single axis.

This change should improve accuracy of computing where exactly the user
swiped.